Added the option to connect 2 edges in the vascular cylinder object, Options to 'save' or 'save as' the modified file, Option to tag the line object or the polygon object.

+++ BrainQuarter.h
@@ -40,7 +40,7 @@ public:
         virtual QSize minimumSizeHint(void) const;
         void getPickedPoint(SoPickedPoint *point,SoPickedPoint *point2);
 signals:
-        void pointPicked(int index, int id, int x, int y, int z,SoType objtype);
+        void pointPicked(int index, int id, float x, float y, float z,SoType objtype);
 		void pointNotPicked();
 
 private:
